allow table selection
